Add response body to BadRequest exception
Add response body to the message because in some cases there is plain text instead of JSON and it is not parsed correctly and the Exception message is null in the result.

There is the issue [#107](https://github.com/spatie/dropbox-api/issues/107) in the spatie/dropbox-api, that was closed without proper fix.

Example of the response body:
```
Error in call to API function "files/upload": Your app is not permitted to access this endpoint because it does not have the required scope 'files.content.write'. The owner of the app can enable the scope for the app using the Permissions tab on the App Console.
```